hello-mcp-practice/pyproject.toml
[project]
name = "hello-mcp-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "mcp>=1.12.4",
    "openai-agents>=0.2.9",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "uvicorn[standard]>=0.35.0",
]
